Move MakeConsensusAndBlock to node/runner/test.go
