+++ lib/node/runner/isaac_simulation_unfreezing_test.go
@@ -8,12 +8,8 @@ package runner
 import (
 	"testing"
 
-	"boscoin.io/sebak/lib/ballot"
 	"boscoin.io/sebak/lib/block"
 	"boscoin.io/sebak/lib/common"
-	"boscoin.io/sebak/lib/node"
-	"boscoin.io/sebak/lib/transaction"
-	"boscoin.io/sebak/lib/voting"
 	"github.com/stretchr/testify/require"
 )
 
@@ -108,50 +104,4 @@ func TestUnfreezingSimulation(t *testing.T) {
 	ba, _ = block.GetBlockAccount(st, kpFrozenAccount.Address())
 
 	require.Equal(t, uint64(ba.Balance), uint64(99999990000))
-}
-
-func MakeConsensusAndBlock(t *testing.T, tx transaction.Transaction, nr *NodeRunner, nodes []*node.LocalNode, proposer *node.LocalNode) (block.Block, error) {
-
-	nr.TransactionPool.Add(tx)
-
-	// Generate proposed ballot in nodeRunner
-	round := uint64(0)
-	_, err := nr.proposeNewBallot(round)
-	require.NoError(t, err)
-
-	b := nr.Consensus().LatestBlock()
-	basis := voting.Basis{
-		Round:     round,
-		Height:    b.Height,
-		BlockHash: b.Hash,
-		TotalTxs:  b.TotalTxs,
-	}
-
-	conf := common.NewConfig()
-
-	// Check that the transaction is in RunningRounds
-
-	ballotSIGN1 := GenerateBallot(proposer, basis, tx, ballot.StateSIGN, nodes[1], conf)
-	err = ReceiveBallot(nr, ballotSIGN1)
-	require.NoError(t, err)
-
-	ballotSIGN2 := GenerateBallot(proposer, basis, tx, ballot.StateSIGN, nodes[2], conf)
-	err = ReceiveBallot(nr, ballotSIGN2)
-	require.NoError(t, err)
-
-	rr := nr.Consensus().RunningRounds[basis.Index()]
-	require.Equal(t, 2, len(rr.Voted[proposer.Address()].GetResult(ballot.StateSIGN)))
-
-	ballotACCEPT1 := GenerateBallot(proposer, basis, tx, ballot.StateACCEPT, nodes[1], conf)
-	err = ReceiveBallot(nr, ballotACCEPT1)
-	require.NoError(t, err)
-
-	ballotACCEPT2 := GenerateBallot(proposer, basis, tx, ballot.StateACCEPT, nodes[2], conf)
-	err = ReceiveBallot(nr, ballotACCEPT2)
-
-	blk := nr.Consensus().LatestBlock()
-
-	require.Equal(t, proposer.Address(), blk.Proposer)
-	require.Equal(t, 1, len(blk.Transactions))
-	return blk, err
 }
